# server/docs/swagger.yaml
definitions:
...
  EngineeringEducationDatabase.PolicyDatabase:
    properties:
      ID:
        description: 主键ID
        type: integer
      acquisitionTime:
        type: string
      additionalRemarks:
        description: 其他备注
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      dataSource:
        description: 数据来源
        type: string
      deletedBy:
        type: integer
      policyContent:
        description: 政策内容
        type: string
      policyCountryOrRegion:
        description: 政策国家或地区
        type: string
      policyName:
        description: 政策名称
        type: string
      policyReleaseDate:
        description: 政策发布时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - dataSource
    - policyContent
    - policyCountryOrRegion
    - policyName
    - policyReleaseDate
    type: object
  EngineeringEducationDatabase.ProfessionalConstructionDatabase:
    properties:
      ID:
        description: 主键ID
        type: integer
      admissionMajor:
        description: 招生专业
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      degreeType:
        description: 授予学位类别
        type: string
      deletedBy:
        type: integer
      isBroadAdmissionCategory:
        description: 是否大类招生
        type: boolean
      isDoubleDegree:
        description: 是否双学位
        type: boolean
      isFirstClassMajor:
        description: 是否一流专业
        type: boolean
      isMajorAccredited:
        description: 是否专业认证
        type: boolean
      isNewEngineeringDiscipline:
        description: 是否新工科
        type: boolean
      isNewMajor:
        description: 是否新专业
        type: boolean
      isSelectedForExcellenceProgram:
        description: 是否入选卓越计划
        type: boolean
      majorCode:
        description: 专业代码
        type: string
      majorName:
        description: 专业名称
        type: string
      majorSatisfaction:
        description: 专业满意度
        type: string
      plannedEnrollment:
        description: 计划招生数
        type: integer
      universityName:
        description: 所在学校
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - admissionMajor
    - degreeType
    - isBroadAdmissionCategory
    - isDoubleDegree
    - isFirstClassMajor
    - isMajorAccredited
    - isNewEngineeringDiscipline
    - isNewMajor
    - isSelectedForExcellenceProgram
    - majorCode
    - majorName
    - majorSatisfaction
    - plannedEnrollment
    - universityName
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.1
paths:
...
  /PCD/createProfessionalConstructionDatabase:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建专业建设库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.ProfessionalConstructionDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建专业建设库
      tags:
      - ProfessionalConstructionDatabase
  /PCD/deleteProfessionalConstructionDatabase:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除专业建设库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.ProfessionalConstructionDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除专业建设库
      tags:
      - ProfessionalConstructionDatabase
  /PCD/deleteProfessionalConstructionDatabaseByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除专业建设库
      tags:
      - ProfessionalConstructionDatabase
  /PCD/findProfessionalConstructionDatabase:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 招生专业
        in: query
        name: admissionMajor
        required: true
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - description: 授予学位类别
        in: query
        name: degreeType
        required: true
        type: string
      - in: query
        name: deletedBy
        type: integer
      - description: 是否大类招生
        in: query
        name: isBroadAdmissionCategory
        required: true
        type: boolean
      - description: 是否双学位
        in: query
        name: isDoubleDegree
        required: true
        type: boolean
      - description: 是否一流专业
        in: query
        name: isFirstClassMajor
        required: true
        type: boolean
      - description: 是否专业认证
        in: query
        name: isMajorAccredited
        required: true
        type: boolean
      - description: 是否新工科
        in: query
        name: isNewEngineeringDiscipline
        required: true
        type: boolean
      - description: 是否新专业
        in: query
        name: isNewMajor
        required: true
        type: boolean
      - description: 是否入选卓越计划
        in: query
        name: isSelectedForExcellenceProgram
        required: true
        type: boolean
      - description: 专业代码
        in: query
        name: majorCode
        required: true
        type: string
      - description: 专业名称
        in: query
        name: majorName
        required: true
        type: string
      - description: 专业满意度
        in: query
        name: majorSatisfaction
        required: true
        type: string
      - description: 计划招生数
        in: query
        name: plannedEnrollment
        required: true
        type: integer
      - description: 所在学校
        in: query
        name: universityName
        required: true
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询专业建设库
      tags:
      - ProfessionalConstructionDatabase
  /PCD/getProfessionalConstructionDatabaseList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: admissionMajor
        type: string
      - in: query
        name: degreeType
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: isBroadAdmissionCategory
        type: boolean
      - in: query
        name: isDoubleDegree
        type: boolean
      - in: query
        name: isFirstClassMajor
        type: boolean
      - in: query
        name: isMajorAccredited
        type: boolean
      - in: query
        name: isNewEngineeringDiscipline
        type: boolean
      - in: query
        name: isNewMajor
        type: boolean
      - in: query
        name: isSelectedForExcellenceProgram
        type: boolean
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: majorCode
        type: string
      - in: query
        name: majorName
        type: string
      - in: query
        name: majorSatisfaction
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: sort
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: universityName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取专业建设库列表
      tags:
      - ProfessionalConstructionDatabase
  /PCD/updateProfessionalConstructionDatabase:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新专业建设库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.ProfessionalConstructionDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新专业建设库
      tags:
      - ProfessionalConstructionDatabase
  /PD/createPolicyDatabase:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建政策数据库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.PolicyDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建政策数据库
      tags:
      - PolicyDatabase
  /PD/deletePolicyDatabase:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除政策数据库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.PolicyDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除政策数据库
      tags:
      - PolicyDatabase
  /PD/deletePolicyDatabaseByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除政策数据库
      tags:
      - PolicyDatabase
  /PD/findPolicyDatabase:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - in: query
        name: acquisitionTime
        type: string
      - description: 其他备注
        in: query
        name: additionalRemarks
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - description: 数据来源
        in: query
        name: dataSource
        required: true
        type: string
      - in: query
        name: deletedBy
        type: integer
      - description: 政策内容
        in: query
        name: policyContent
        required: true
        type: string
      - description: 政策国家或地区
        in: query
        name: policyCountryOrRegion
        required: true
        type: string
      - description: 政策名称
        in: query
        name: policyName
        required: true
        type: string
      - description: 政策发布时间
        in: query
        name: policyReleaseDate
        required: true
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询政策数据库
      tags:
      - PolicyDatabase
  /PD/getPolicyDatabaseList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: endPolicyReleaseDate
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: policyCountryOrRegion
        type: string
      - in: query
        name: sort
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: startPolicyReleaseDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取政策数据库列表
      tags:
      - PolicyDatabase
  /PD/updatePolicyDatabase:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新政策数据库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.PolicyDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新政策数据库
      tags:
      - PolicyDatabase
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"